backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
passlib==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0